refactor CompactorManager
