+++ src/meta/src/hummock/manager/mod.rs
@@ -884,7 +884,8 @@ where
         compact_status.report_compact_task(compact_task);
         if let Some(context_id) = context_id {
             self.compactor_manager
-                .report_compact_task(context_id, compact_task);
+                .report_compact_task(context_id, compact_task)
+                .await;
         }
         let task_status = compact_task.task_status();
         debug_assert!(
@@ -1043,7 +1044,7 @@ where
                         return;
                     }
                 };
-                let compactor = match self.compactor_manager.random_compactor(None) {
+                let compactor = match self.compactor_manager.random_compactor(None).await {
                     None => {
                         tracing::warn!(
                             "Skip committed SST sanity check due to no available worker"
@@ -1414,7 +1415,10 @@ where
         };
 
         // 2. select_compactor
-        let compactor = self.compactor_manager.random_compactor(Some(&compact_task));
+        let compactor = self
+            .compactor_manager
+            .random_compactor(Some(&compact_task))
+            .await;
         let compactor = match compactor {
             None => {
                 tracing::warn!("trigger_manual_compaction No compactor is available.");
