+++ src/meta/src/rpc/service/hummock_service.rs
@@ -189,7 +189,8 @@ where
             .await?;
         let rx = self
             .compactor_manager
-            .add_compactor(context_id, req.max_concurrent_task_number);
+            .add_compactor(context_id, req.max_concurrent_task_number)
+            .await;
         Ok(Response::new(RwReceiverStream::new(rx)))
     }
